requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.0
python-decouple>=3.8
PyMySQL>=1.1.0
cryptography>=41.0.0
web3>=6.11.0
celery>=5.3.0
redis>=5.0.0
pillow>=10.0.0
django-filter>=23.0
psutil>=5.9.0
requests>=2.31.0
python-dateutil>=2.8.0
